Create wallets for users when wallet ID field is empty
